refactor: improve issue context structure
