+++ user-application/src/main/resources/io/github/dronesecurity/userapplication/presentation/reporting/issue/newIssue.fxml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.TextArea?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.layout.AnchorPane?>
+
+<AnchorPane xmlns="http://javafx.com/javafx/11" xmlns:fx="http://javafx.com/fxml/1" prefWidth="450.0" prefHeight="300.0" fx:controller="io.github.dronesecurity.userapplication.presentation.reporting.issue.NewIssueUIController">
+    <Label layoutX="14.0" layoutY="21.0" text="Subject" />
+    <TextField fx:id="issueSubject" layoutX="81.0" layoutY="17.0" prefHeight="25.0" prefWidth="315.0" />
+    <Label layoutX="14.0" layoutY="54.0" text="Details" />
+    <TextArea fx:id="infoTextArea" layoutX="14.0" layoutY="80.0" prefHeight="166.0" prefWidth="382.0" wrapText="true" />
+    <Button layoutX="367.0" layoutY="253.0" mnemonicParsing="false" onAction="#sendIssue" text="Send" />
+   <padding>
+      <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
+   </padding>
+</AnchorPane>
